requirements.txt
requests>=2.31.0
newspaper3k>=0.2.8
readability-lxml>=0.8.1
beautifulsoup4>=4.12.0
wayback-machine-scraper>=0.1.0